requirements.txt
moderngl>=5.8.0
pygame>=2.5.0
numpy>=1.24.0
numba>=0.58.0
PyGLM>=2.7.0
Pillow>=10.0.0
PyWavefront>=1.3.3
pytest>=7.4.0
pytest-cov>=4.1.0
